requirements.txt
PyYAML==5.4
loguru==0.7.0